Fix naming, add export interface
