+++ addons/verlet_rope_4/Data/RopeParticleData.cs
@@ -67,13 +67,27 @@ public sealed class RopeParticleData
     }
 }
 
+// TODO: Make providers implement interface/base with only PositionCurrent, so that bookmark properties were not exposed when not needed
 public struct RopeParticle
 {
+    /// <summary> The current position of the particle for this frame - used for mesh generation. </summary>
     public Vector3 PositionPrevious { get; set; }
+
+    /// <summary> Bookmark provider property - The position of the particle from the previous frame. used to calculate velocity. </summary>
     public Vector3 PositionCurrent { get; set; }
+
+    /// <summary> Bookmark provider property - The acceleration applied to this particle (i.e. combined from gravity, wind or any other forces). </summary>
     public Vector3 Acceleration { get; set; }
+    
+    /// <summary> Bookmark provider property - Indicates whether particle's position is locked and not simulated (e.g. for attachment points). </summary>
     public bool IsAttached { get; set; }
+    
+    /// <summary> Internal property - Provides currently calculated visual tangent particle vector. </summary>
     public Vector3 Tangent { get; set; }
+    
+    /// <summary> Internal property - Provides currently calculated visual normal particle vector. </summary>
     public Vector3 Normal { get; set; }
+    
+    /// <summary> Internal property - Provides currently calculated visual binormal particle vector. </summary>
     public Vector3 Binormal { get; set; }
 }
