+++ addons/verlet_rope_4/IVerletExported.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace VerletRope4;
+
+public interface IVerletExported
+{
+    public static string ScriptPath => throw new NotImplementedException();
+    public static string IconPath => throw new NotImplementedException();
+    public static string ExportedBase => throw new NotImplementedException();
+    public static string ExportedType => throw new NotImplementedException();
+}
