+++ addons/verlet_rope_4/Rendering/VerletRopeMesh.cs
@@ -5,8 +5,13 @@ using VerletRope4.Data;
 namespace VerletRope4.Rendering;
 
 [Tool]
-public partial class VerletRopeMesh : MeshInstance3D
+public partial class VerletRopeMesh : MeshInstance3D, IVerletExported
 {
+    public static string ScriptPath => "res://addons/verlet_rope_4/Rendering/VerletRopeMesh.cs";
+    public static string IconPath => "res://addons/verlet_rope_4/icon_rope_mesh.svg";
+    public static string ExportedBase => nameof(MeshInstance3D);
+    public static string ExportedType => nameof(VerletRopeMesh);
+
     private const string DefaultMaterialPath = "res://addons/verlet_rope_4/Materials/rope_default.material";
     private const string CreationStampMeta = "verlet_rope_internal_stamp";
